+++ Thunderstore/CHANGELOG.md
@@ -56,3 +56,10 @@
 
 -Remove server calls temporarily. LethalNetworkAPI appears to no longer work with vanilla.
 
+## 1.9.0
+
+-Fixes sync, allowing host to sync gameplay settings with players.
+-Adds audio effects.
+-Adds experimental emergency teleport feature. Teleports all players back to ship. Can be used once per day. Use with caution.
+
+